Add control over cue label font, and display of labels.
Also cleaned up a bunch of IntelliJ warnings due to the fact this is a
library, not a standalone application.

+++ src/main/java/org/deepsymmetry/beatlink/data/ArtFinder.java
@@ -362,7 +362,7 @@ public class ArtFinder extends LifecycleParticipant {
 
         // Send the artwork request
         Message response = client.simpleRequest(Message.KnownType.ALBUM_ART_REQ, Message.KnownType.ALBUM_ART,
-                client.buildRMST(Message.MenuIdentifier.DATA, slot.slot, trackType), new NumberField((long)artworkId));
+                client.buildRMST(Message.MenuIdentifier.DATA, slot.slot, trackType), new NumberField(artworkId));
 
         // Create an image from the response bytes
         return new AlbumArt(new DataReference(slot, artworkId), ((BinaryField)response.arguments.get(3)).getValue());
